pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "flask-dance[sqla]>=7.1.0",
    "flask-login>=0.6.3",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "oauthlib>=3.3.1",
    "openai>=2.9.0",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.10.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.44",
    "stripe>=14.0.1",
    "weasyprint>=66.0",
    "werkzeug>=3.1.3",
]
